U5: vault store layout + versioned header + crash-safe file discipline — the D42 partition made structural (a closed three-variant BesideFile enum: config.toml, vault.header, vault.lock; everything else resolves under the ciphertext-only store/ area), a frozen header envelope (magic + [version, body-bstr]) whose v1 body carries an opaque size-capped KDF slot for U6 and the D50 wrap-mode registry, atomic temp+fsync+rename writes with a deterministic kill matrix, and an flock-backed unlink-free single-writer lock
Header: deterministic CBOR via antseal-core's strict codec (project rule 5;
no new dependency); the envelope is frozen forever so ANY build reads the
version and answers 'vault created by newer antseal' cleanly (mapped to the
U2 vault-newer-version class, code 16, without parsing the future body);
constructors enforce the decode-side caps (the F41 constructible==decodable
lesson); golden vector committed as an in-test literal (NON-SECRET filler —
no real KDF params exist at U5) + 12-row tamper matrix incl. truncation
sweep, non-canonical ints, arity/key violations, cap bombs. The exact
encode() bytes are the AAD U6 binds (D42 rider recorded in module docs).

Atomic write: same-dir temp -> write -> fsync -> rename -> parent-dir
fsync; failpoint-injected kills at every stage prove old-or-new-never-torn
plus recovery, and real errors clean their temp up (leftovers from kills
are inert dotfiles). Lock: std File::try_lock (stabilized 1.89; probed on
the pinned 1.92) — the kernel releases a dead holder's lock, so stale
lockfiles are structurally inert (no pid-liveness heuristics) and the file
is never unlinked (inode-swap race documented in module docs). Same-process
double-acquire refused (per-OFD semantics, probed). Vault dir: ~/.antseal
default, ANTSEAL_DIR env override (a path, not a secret — D41-compatible;
pure resolve_from() core keeps tests set_var-free under edition 2024).

Scoped per wave plan: NO KDF/AEAD here (U6 next wave); the record-splice
AAD test lands with U6/U9 where an AEAD first exists.

+++ crates/antseal-cli/src/error.rs
@@ -408,7 +408,7 @@ pub enum CliError {
         "this vault was created by a newer antseal (vault format v{found}; this build \
          supports up to v{supported}): upgrade antseal instead of downgrading the vault"
     )]
-    VaultNewerVersion { found: u32, supported: u32 },
+    VaultNewerVersion { found: u64, supported: u32 },
 
     /// Distinct from gas by spec (core flow 1): the ANT token balance
     /// cannot cover the quote.
@@ -511,7 +511,7 @@ pub enum CliError {
         "this export file was written by a newer antseal (export format v{found}; this \
          build supports up to v{supported}): upgrade antseal to import it"
     )]
-    ImportNewerVersion { found: u32, supported: u32 },
+    ImportNewerVersion { found: u64, supported: u32 },
 }
 
 impl CliError {
